+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "qstash-queue",
+  "name": "dalle-2-generator",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -19,6 +19,8 @@
     "swr": "^1.3.0"
   },
   "devDependencies": {
+    "@types/node": "18.11.13",
+    "@types/react": "18.0.26",
     "autoprefixer": "^10.4.13",
     "eslint": "7.32.0",
     "eslint-config-next": "13.0.0",
