+++ pages/api/image.ts
@@ -1,8 +1,12 @@
+import { NextApiRequest, NextApiResponse } from "next";
 const QSTASH = `https://qstash.upstash.io/v1/publish/`;
 const DALL_E = "https://api.openai.com/v1/images/generations";
 const VERCEL_URL = "https://dalle-2.vercel.app";
 
-export default async function handler(req, res) {
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
   const { prompt } = req.query;
   try {
     const response = await fetch(`${QSTASH + DALL_E}`, {
@@ -22,7 +26,7 @@ export default async function handler(req, res) {
     const json = await response.json();
     return res.status(202).json({ id: json.messageId });
   } catch (error) {
-    res
+    return res
       .status(500)
       .json({ message: error.message, type: "Internal server error" });
   }
