+++ utils/use-interval.tsx
@@ -1,7 +1,7 @@
-import React, { useState, useEffect, useRef } from "react";
+import { useEffect, useRef } from "react";
 
-export const useInterval = (callback, delay) => {
-  const savedCallback = useRef();
+export const useInterval = (callback: () => void, delay: number | null) => {
+  const savedCallback = useRef<any>(null);
 
   useEffect(() => {
     savedCallback.current = callback;
